move meetup logic into its own directory #33
